refacto: se genera proyecto common
Se genera el proyecto common para poder hacer uso de utilidades compartidas
