+++ NominaDownloaderPEIGTO.sln
@@ -17,6 +17,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0AB3BF05
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NominaDownloaderPEIGTO.Tests", "tests\NominaDownloaderPEIGTO.Tests\NominaDownloaderPEIGTO.Tests.csproj", "{3F498D3D-B34C-448C-84F1-39F198C0B1D1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NominaDownloaderPEIGTO.Common", "src\NominaDownloaderPEIGTO.Common\NominaDownloaderPEIGTO.Common.csproj", "{D949A6CE-E065-4604-84F2-1A607F5DA104}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -87,6 +89,18 @@ Global
 		{3F498D3D-B34C-448C-84F1-39F198C0B1D1}.Release|x64.Build.0 = Release|Any CPU
 		{3F498D3D-B34C-448C-84F1-39F198C0B1D1}.Release|x86.ActiveCfg = Release|Any CPU
 		{3F498D3D-B34C-448C-84F1-39F198C0B1D1}.Release|x86.Build.0 = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|x64.Build.0 = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Debug|x86.Build.0 = Debug|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|x64.ActiveCfg = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|x64.Build.0 = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|x86.ActiveCfg = Release|Any CPU
+		{D949A6CE-E065-4604-84F2-1A607F5DA104}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -97,5 +111,6 @@ Global
 		{A86989BC-6F40-4531-99B3-49963D29748B} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{2B0729AB-7194-4E44-B0D3-477DD5D0DB4D} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{3F498D3D-B34C-448C-84F1-39F198C0B1D1} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{D949A6CE-E065-4604-84F2-1A607F5DA104} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
